fix: make jwt negative test message more informative
The test for invalid key is a negative test, actually it should
warn when an invalid signature is passed. But the log message is
not clear and gives the user the sensation that tests are failing.

Add a more informative message to be clear that this test is
correctly passing

Closes: #127