feat: add traces support using opentelemetry
Add tracing to grpc interceptor so we can start adding
observability to the application. It currently defines a
trace exporter based in stdout, but could be configurable in
the future.

Config allows to enable or disable traces, and adjust the ratio
for them (typically set to 0.1). Added some examples of subtraces
to monitor conflicting/crucial parts of the code.

Closes: #154