feat: add initial version of auth middleware
Add a new auth interceptor to the grpc server. Initially it
does a basic check for endpoints that need to bypass auth, and
the ones needing it.
For the ones needing auth, it checks that the token is valid,
and extract the claims for the user.
Also add support for passing credentials on the cli side.
More PRs need to come to add proper ACL

Partially-closes: #12