Implement New Enrollment Flow
This change allows the gathering of an OAuth2 token.

An approach is taken so that the state (nonce) and port
are recorded to the database which allows a mapping from
the group_id to the token.

Token is encrypted into the DB

After this pull request, the webhook register and webhook
listener will be pushed. Broke these into two patches to help
make review easier.

Signed-off-by: Luke Hinds <luke@stacklok.com>