Add complete authorization workflow (#233)
* feat: add initial version of auth middleware

Add a new auth interceptor to the grpc server. Initially it
does a basic check for endpoints that need to bypass auth, and
the ones needing it.
For the ones needing auth, it checks that the token is valid,
and extract the claims for the user.
Also add support for passing credentials on the cli side.
More PRs need to come to add proper ACL

Partially-closes: #12

* feat: populate user claims as part of jwt token

In order to facilitate authorization process, it is
desirable to have all the data related to the user
stored in the jwt token. This needs to be followed by
a proper way to revoke tokens when there is some modification
of the user

Partially-closes: #12

* add layer for validating superadmin methods

* fixes in auth

* add validation based on method and request

* add complete authz workflow

* fix tests

* fix problems with rebase

* add authorization tests