feat: use different workflow for closing browser
Closing a browser when a shutdown method is received is not
going to work, because server and client are not in the same
network, and is likely that client cannot receive commands
as the IP is not being exposed.
Instead, create a coroutine that checks regularly until the
token is created, and close when that happens.

Closes: #283