feat: add make target to generate db schema
Add a target that spins up a postgres, and an schema spy container
and applies migration to the db. It uses schemaspy to generate
a png image, then copies it to documentation.

schemaspy generates additional documentation that could be used
as well, but for the moment we just copy the schema diagram

Closes: #158