Make database migrate up idempotent
If there is no change in the database migration, we don't need to fail the
command. The reasoning behind this is that we could add the migration command
to the docker-compose file (or a kube manifest if we ever get one), and it could
simply run the migration with the safe assumption that it'll pass.

Note that the fact that it's idempotent doesn't mean it can be ran in parallel.