Add database migration to docker-compose
This enables the database migration step to the docker-compose definition. We can
thus now start development with the following:

```bash
$ docker-compose -d migrate
```

This will spawn the postgres database, and once it's healthy, it'll run the
migration command.

I also had to modify the mediator container calls to remove the `mediator-server`
command, since we already have that in the container's entrypoint.