Implement DB sync query and update repositories schema
This function will be called after intial enrollment and is a
precursor to calling the RegisterWebhook function.

The idea is that we populate the database with all repos.

We then provide a handler to queries the database and return entries
to the client. The client can then select from the list and 'register'
those repos with a webhook

The function can also be called at a determined stage to ensure
mediators database is synced to the current repos in the users
github account.

This is part of the overall implementation tracked in #100

Signed-off-by: Luke Hinds <luke@stacklok.com>