fix: do not make group id mandatory for provider enrollment
When we ask to enroll into a provider, we need to know the group
that needs to be enrolled, so we made this field mandatory.
However, an usual case can be that an user only belongs to a
group, so we actually can relax the requirements there, and try
to infer the group id from the claims, if there is only one.

Closes: #303