Add metrics support
Through the usage of the `metrics` section in the configuration, we may now use
OTel-powered metrics!

This uses prometheus as the default exporter for now, but we can make this
configurable in the future.

To enable, the settings `metrics.enabled` needs to be set to `true`, in a
similar fashion than the tracing configuration.

The easiest way to test this is as follows:

```bash
docker-compose -d migrate
go run ./cmd/server/main.go serve &
curl 127.0.0.1:8080/metrics
```