feat: allow to use transactions when several queries are involved
There are several operations that involve creation of several
resources in the db that need to be atomic. If there is a failure
in any of the steps, the whole transaction needs to be rolled back.
So we need to use transactions for it. Expose the functionality
in store interface, and use it for the needed methods.
Also fix some issues with unique indexes in the db.

Closes: #321