Migrate container to use scratch as a base and non-root user
To decrease the attack vector and follow hardened defaults, this changes
the base image to be `scratch` with a `nobody` user that's set up
in the container.

Since our defaults actually work out of the box, this also removes the
mounting of the config. We now only rely on the minimal the containers need.