Build and release container
This sets up a github job that will build the container. In cases where the
job comes from a PR, it'll only build it. In cases where it's a push to `main`
it'll push a `latest`, which allows us to get a new build for everything
that goes to `main`.

Finally, in cases where it's a tag, an appropriate container tag matching
the git tag will be generated, which we could use in a release as needed.

This uses GHCR to host the container.

In order to ensure that we don't attempt writing packages on pull request
this creates a new workflow file for pull requests. proper permissions were
given to the main and tags workflows.