Change reference to default database
In the docker-compose file we were already setting up the default database
as `mediator`. It might have been by mistake, but we had been using the
`postgres` database in our configuration. This fixes that; thus, the
database we reference in our config is now `mediator` which is a little
more suiting to the project.

Note that you'll need to update your references in case you were using
external tooling to develop (such as DBeaver).