fix: include provider in repo registration
The same as we ask for provider when we enroll the repos,
we should be asking for it when we register the repos.
Added that on grpc methods, and on client.
I also made group-id optional, trying to infer that from the
claims as it was done on the enroll command, so it behaves
in the same way.

Closes: #382