Centralize configuration
This adds a central structure for our configuration under `internal/config`.

The idea is that this is where all the settings will be placed and registered
so we can have an easier way finding and bootstrapping new settings as the
application grows.

To begin with, the HTTP and GRPC settings were migrated.