Initial implementation of centralized eventing
This adds wrappers over Watermill so we can easily configure, use
and add subscribers for the events we care for.

The idea is to have a Registrar per provider which will contain the
events, and when the Eventer is created, the Registrar will walk
through the handlers it has available and request messages from them.