Implement get and list of repositories
This implements a new endpoint to list a single repository and get further information about it.

On the other hand, this also implements a `list` and a `get` sub-command so we can both
know the state of our current repositories and get extra info on individual ones.

It adds a simple `registered` flag that tells us if the repo has been registered in mediator (for usability).

Finally, this also had to roll back the sqlc change since it breaks our sql generation.

Closes: #433