CI: Add CI test to verify that sqlc has been run
This sets up a simple CI job that runs sqlc and verifies that it's been
run recently and it works as expected.

The idea is that we shouldn't accept PRs that haven't had sqlc auto-generate
the SQL queries. This also helps us verify that new changes to sqlc don't
break the code generation.