ci: Remove proto-breaking changes check
We had tried to continue on error but this is not cutting it as changes are
still appearing as red. We are not in a stage where we have a stable API, so
let's not do these types of checks just yet. Instead, let's defer checking this
until we declar the API stable.