ci: Make sqlc CI job more general and test for mockgen changes
This adds a check for uncommited `make mock` changes on CI. Thus ensuring that
folks don't merge code without updating the mocks as needed.

Note that this was added to the same job that runs the sqlc checks because
it would run `make bootstrap` twice otherwise, which is time-taking and wasteful.