Create database schema for projects
This presents the database schema and relevant queries for representing
projects as a directory tree.

There is one `Root Project` created by default.

The intent is that users will:

* Create projects under a given parent.
* List the sub-projects they have access to.

We also have queries for getting the parents of a project, however,
these won't be exposed to users and will be used for internal purposes
such as authorization constructs.