feat: allow to specify owner when enrolling a provider
github doesn't provide a clear way to generate specific tokens
just tied to an org, and does not have any api to get if the token
is for user or org. There is no support for that on pat tokens,
and very limited for fine grained user tokens.

So instead an easier way to do it, is to just pass a --owner flag
when enrolling the provider. If not set, it assumes that we want to
just get the user resources. If set, it gets resources just for the
org specified in the filter.

For example:

- medic provider enroll --provider github -> will return my repos under yrobla
- medic provider enroll --provider github --owner yrobla-org -> will return my repos under yrobla-org (providing i have access to them)

Also refactor code to add the same logic for oauth and pat tokens, and move to repositories

Closes: #585