repo_register: A more user-friendly error message if no repos are listed
When no repos could be listed (e.g. an org with only private
repositories was enrolled), registering the repos would fail with a
strange-looking message:

```
Error getting repo selection: please provide options to select from
```

This is because the code would try to run a `survey` on an empty list of
repos. Let's preempt that codepath by providing a nicer error message
sooner, before launching the `survey`.