Standardize handling of entity types around an enum instead of strings
During review of the previous patch that was adding the `status get`
command it was pointed out that many of our commands and the
request/response grpc structs and messages alike presume that the entity passed
around is a repository, because that was the initial type.

At the same time, handling of the different types was a bit ad-hoc
across the codebase, some places used just a string, some a string-like
type alias.

This patch introduces a new enum in the proto definition called
EntityType. This enum replaces the use of string types in messages that
deal with entity types, making those strongly typed.

There was an existing Entity enum that was only used in a provider which
I kept, but moved into the Provider message - I think it's good to have
both because providers can "export" types that mediator does not handle.
This enum was renamed to ProviderEntity and mvoed to the Provider
message.

The other nice thing about the generated enum is that there are maps
converting from enum values to strings and vice versa. This is useful
for converting from CLI, validation and as a way to print all allowed
values. As a side-effect, we can now drop the EntityType string alias.