cli: Use viper.GetInt32("group-id") directly for group-id instead of util.GetConfigValue
We were using util.GetConfigValue to get the group-id and relied on the return
value being `int(32)`, but this assumption doesn't hold, because
`util.GetConfigValue` first calls `viper.Get(key)`, attempting to get the
configuration file key. If the `key` is the same as the command line flag and
the command line flag has a default defined, as it's the case for `group-id`,
then for any numerical types, `viper.Get` would have returned the default
value, typed as an `int`:

```
case int32, int16, int8, int:
    return cast.ToInt(val)
```

which would then trip up the explicit type cast to `int32` we had after
calling `util.GetConfigValue`.

Because we're really only interested in the command line flag in this
case, let's not use `util.GetConfigValue` and confuse us with type cases
back, but just straight use `viper.GetInt32`.

In general, `util.GetConfigValue` should be used to retrieve config
values optionally overriden by command line flags, but for values that
are only defined with command line flags, especially numerical.

In addition, it is odd to have the default defined both when declaring
the flag and when retrieving the value.