db: Add nullable and properly coalesced artifact_id
In addition to referencing repository entities from the
rule_evaluation_status table, we need to reference artifacts as well.

To do that, we add a new column artifact_id to the
rule_evaluation_table so that we can take advantage of cascaded deletes.

Because this column can be nullable, we need to COALESCE() the column
to a valid integer value, but one that's not valid for the model because
otherwise comparing NULL that's passed in with NULL in the conflicting
row never yields true. The same COALESCE() must be applied to the unique
index.