util: Use dedicated test package for unit tests
This uses a dedicated test package as opposed to including tests as part
of the same `util` package.

The idea with this pattern is that it allows us to test how a regular
test implementation would actually use the functions. It also, deters
us from testing specifics within the implementations which is not ideal.