engine: Handle skipped rules and silent skips
This adds the needed functionality to skip rules from either an ingestor or an
evaluator, as well as silently skip rule evaluation for an event that
doesn't match certain criteria. The main use-case is artifacts that may
trigger a policy run for artifacts that are not relevant to a policy.