protobuf: Use VersionedArtifact, get rid of ArtifactEventPayload
The ArtifactEventPayload protobuf message was a relic of how the
artifacts were initially implemented by just looking at the github API.
It had several problems, mainly that it didn't reflect how artifacts are
stored in the DB, supported only a single tag.

Let's converge on using VersionedArtifact which just contains an
artifact and one version. This way we would be able to use the same
struct in webhooks, init and reconciler.