Add artifact ingester, change the schema of artifact_signature
The schema of the artifact_signature rule_type was built with an
artifact having a single tag and was built to be used through the now
removed ArtifactEventPayload struct. Changing the schema allows us to
specify the tags better, as a subset of container tags.

Since evaluating the schema with the builtin ingester would be too
complex and would require special-casing the individual keys and values,
it also makes sense to switch to a new ingester created specifically for
artifacts.