engine: Add rego evaluation engine
This adds rego as an engine for rules, it takes whatever input
from the ingester and is able to evaluate it against OPA/rego policies.
We also pass in the policy from mediator so we could use that as input
if needed.

There are two modes of operations:
- `deny-by-default`: Which is more secure and expects policy writers to only set an `allow` variable to true if the policy matches expectations
- `constraints`: will deny the request if a violation is found. It can return a message with further explanations for the denial.

the reason why I didn't just use the `deny-by-default` as the only engine is that it's non-trivial to return
meaningful messages to the user with such a model in rego.