Use a generic version of JQ accessor, called JQGetTypedFromAccessor to parse GitHub payload
Addresses a comment brought up during an earlier review that the way we
parse the github payload is unsafe - in case the JQ accessor wouldn't
return anything, the value returned would have been `nil` (the default
value of `any`) which would panic on type assertion. Similarly, if the
assertion failed due to github changing its API.

Because these attributes are all required by the layer calling
`JQGetTypedFromAccessor`, I opted to return an error in case the
accessor doesn't find a value which would be propagated to the caller
and handled appropriately.

Other parts of the mediator where `nil` can be handled more gracefully
(e.g. because it is just passed to `reflect.DeepEqual`) still use the
old function `JQGetTypedFromAccessor` which returns `any`

Fixes: #785