containers: Suppress error message when trying to fetch signatures, split extractAndValidateSignature into a separate function
When fetching container signatures, we attempt to do that for all
container versions and their digests we receive. In case the container
is not signed, we would throw a scary-looking message. The error itself
is handled well in the code by returning an empty signature object, we should
just make the message in the logs more friendly than:

```
{"Timestamp":1693470050051776000,"message":"2023/08/31 10:20:50 error
getting manifest: error getting image: GET
https://ghcr.io/v2/jakubtestorg/testrepo/manifests/sha256-729d8897773756c0c84f785aa2b44f05174fdf19de12a4497ecec042ecd31934.sig:
MANIFEST_UNKNOWN: manifest unknown"}
```

Also splits out the extraction and validation of a signature into
its own function.