Putting it all together: Enables us to run rego rules on git contents
This adds the pieces together that we need in order to run rego rules on git content.

It proposes a new example rule that verifies that we have codeQL in a repository
to do static analysis.

In order to get this working, we had to create two custom rego rules that execute
on the memoryfs that we got from the git ingester.

There were two bugs that were fixes with this:
* The `clone_url` was reset when registering the webhook.
* We were casting to the wrong type in the git ingester.