refactor: Move webhook event parsinglogic to webhook handler
This moves the event parsing logic to the webhook handler since that's
already github specific. Instead of doing any JSON mangling on the executor
side, it will expect already built events that are ready to be processed.

This makes the executor lighter and allows us to add more
complex logic on the webhook side (e.g. handling pull requests)