cleanup: remove unnecessary checkups from rule
The rule previously known as "actions_permissions" was a little complex in
the sense that it checked a couple of attributes from the github repo
configuration.

This renames the rule to something more appropriate and checks for only one relevant
thing. Thus making it more understandable.

This also adds appropriate guidance for folks.