medev: a CLI tool to work with mediator
The intention is to separate development (e.g. rule type development) tools from
the day-to-day CLI utilities. This makes the `medic` CLI slimmer and allows us to pack
more dependencies and functionality into the new `medev` command.

For now, this merely separates the code, subsequent PRs will deal with releasing a
binary.