cli: Add table output for `policy_status` sub-command
This adds table output for the `policy_status` subcommand, making the
output a little more friendly. It also adds colors and emojis to the status
column as to reflect the reader's eyes towards that.

The `--all` flag in the `policy_status list` was changed to `--detailed` to
reflect better what the flag does.