rego: Add option to skip for `deny-by-default` evaluator (#903)
There are cases where a policy is not applicable for an entity. e.g. We shouldn't check for npm dependencies on a golang repository.

This PR allows us to skip a rule evaluation in the rego engine if we mark a `skip` expression in the rego policy.

An example was added to the dependabot rule. We may skip the rule if a certain dependency file doesn't exist.

The rule was renamed to `dependabot_configured` to reflect the name, and the policy may now take paramters that
define the package ecosystem, interval schedule, and finally the file that should exist for the rule to apply.