Sort CSV list of known entities to prevent needless doc updates
We return a comma-separated list of known entities which is used to
generate the known entities in error messages and help output, but it
turns out that the order is not stable because we read the known
entities from iterating over a map. Let's sort the keys before returning
them to prevent getting documentation update PRs that just switch the
order.