fix/refactor: Move `auth.token_key` and other parameters to `AuthConfig` structure
The intent is to have a central place to fetch the configuration and validate it.

This way, starting the server will fail if the needed crypto configuration is not
set up appropriately.

This also moves the `crypto` package to `internal` which is more appropriate.

Finally, the concept of `auth.token_key` changed from being a string to being a file.
The idea is that we'll have all secrets as files which will be referenced as kubernetes secrets.

Closes: #923