Simplify policy structure and rename
Only me and my dog understand why the policy had the concept of contextual rule sets.

This is not a feature we're actually using, and makes explaining these concepts to folks
quite confusing.

So... let's get rid of pluggable contexts within a policy and let's just declare the provider
at the top of the policy file. This way it's more simple and covers the use-cases we're currently working on.

When we do need another provider for a policy, we can then just create a separate policy file and
manage it that way.

While we're at it, this also renames `PipelinePolicy` to simply `Policy`.