vulnerability evaluator: Submit reviews with just comment if the author is the same as the mediator identity
Because github doesn't let you review your own code, in case mediator is
running as the same identity as the contributor submitting the PR, we
can't request changes.

In that case, let's just mark the review as commenting instead of
changes requested. A follow-up patch would then set the commit status to
failed to prevent merge.

Fixes: #951