Add initial introduction of providers to the database
This simply replaces the usages of the `github` provider checks and instead
provides a simple implementation of providers that will be expanded upon.

The intent is to leave the same functionality as-is, but instead rely on a
provider that's installed on every organization by default.

I've left traces of how the implementation will evolve by adding a definition
and a `providers_type` column to the providers database. These will serve as the means
of instantiating providers dynamically, instead of always building the github one
as we do today. But, to keep things short, I've left that out of this PR.