Extend PR vulnerability checks with a configurable action to set commit status
Because using code reviews to block commits might be problematic for
self-enrollment, this patch adds a new `action` enum value named
`commit_status`. When the PR vulnerability check is enabled and this
action is selected, mediator will either mark the commit status as
succeeded if no vulns are found or failed if vulnerabilities are found.

In the repository branch protection rules, the user can then configure a
rule that would require the `mediator.stacklok.dev/pr-vulncheck` rule to
pass.

The vulnerable dependencies are still pointed out in review comments,
but the review is submitted as "COMMENT", the blocking decision is made
by the commit status.

Fixes: #935