Actual providers implementation
This defines the provider implementations in such a way that we
can instantiate them and check if a rule type is able to accomodate for those.

This is done through a new `ProviderBuidler` implementation which is in
charge of instantiating the provider type that's needed at a given time.

Note that we still use the GitHub provider for HTTP calls since it has a lot
of handy features (such as rate limiting handling) that we could leverage.