vulncheck: Support Go depenndecies in PR vulnerability scanner
Builds atop PR #1012 by supporting lookup of Go dependencies in OVS as
well as their checksums in the Go sum database.

Because we need to do another lookup in the go sum database, we also
need to keep state during the lookups, so I collapsed the two methods of
the RepoQuerier interface into one.

Fixes: #912