feat: Persist provider configuration in protobuf
This moves the provider configuration from the providers package into the protobuf definitions.

This enables us to be able to configure/create/update providers via protobuf in the future.

For now, we only have the protobuf structure, but in the future we'll have a `ProvidersService`
API that we'll be able to leverage in order to manage providers.