providers: Change the signature of NewRequest to accept body as interface{}
The providers.REST interface defined a NewRequest() method that was
implemented by the generic HTTP REST provider as well as the github REST
provider. However, the two implementations treat the body of the request
a bit differently - the HTTP REST provider treats the body as a generic
`io.Reader` which was also exposed in the high level interface, but the
github NewRequest treats the body as an interface that can be
JSON-encoded.

And because io.Reader doesn't really encode to JSON easily, but [expects to be
able to infer the types](https://pkg.go.dev/encoding/json#Marshal) from the
interface it's encoding, I think it's better to just expose `interface{}` in
the high-level provider interface and let the lower-level implementation deal
with the conversion. This way the caller would know to pass e.g.
`map[string]any` instead of plain io.Reader.

If we want to keep the reader in the high level interface, then another
option might be to encode the io.Reader into map[string]any in the
`RestClient.NewRequest` method.