cleanup: Use projects instead of groups and orgs
This takes into use the concept of projects in Mediator, thus replacing our usage
of groups.

In the background, we also got rid of the organizations SQL table and use projects
underneath. Organizations are identifiable via a `is_organization` boolean in the table.

Projects and Organizations now use a UUID index.