cleanup: Allow for policy files to not specify projects
This makes it easier to create policies for self-registered users, since
they can simply do `medic policy create -f <path to file>` and mediator
will just adopt the default project for the user.

This also moves the policy validations to the `validations.go` file in our
API, since it makes things more predictable.

Finally, this removes the context validation, as it's not super useful and we anyway
check for this along the code-base.