fix: Track rule type instantiations in profiles
This starts tracking rule type instantions in profiles by introducing a table to track
these. This table is populated when creating a new profile and is checked
before deleting a rule type. We thus not allow folks to delete rule types
if there is a profile instantiating it.

Note that the rule type IDs and profile IDs are tracked in the `entity_profile_rules`
table using a the `ON DELETE CASCADE` feature. Thus, when a profile is deleted, the
instantiations get cleaned up as well.

Note that we currently don't support updated profiles (yet), so there was nowhere to
add this logic for such a case. This will need to be considered once that feature is
implemented.

Closes: https://github.com/stacklok/mediator/issues/1112