tests: Add env variable to indicate that the database instance shouldn't be cleaned up
If the `MEDIATOR_TEST_DONT_CLEAN_DB` is detected as set, this prevents the in-process-managed DB to be
cleaned up.

This is useful when needing to verify the state of the db after the test failed.

To find the process you'd need to do `ps aux | grep postgres` and kill it eventually.