harden: Only store registered repositories
This changes the logic of the repo register to only store registered repositories
as opposed to registering everything and then verifying if they're registered or not.

This also uncovered that we weren't really checking for issues in registration... So,
this changed that logic.

This also adds a new provider called `repo-lister` which enables implementors to list
repositories. This is handy for the auto-enrollment flow, which requires us to list remote
repositories, compare them to the registered ones, and then enroll.

We might not have auto-enrollment (webhook creation) for other providers... but it would be
handy to be able to list repos so we could then build requests that manually register the repos
to mediator.