Remove domain model (part 1 CLI)
One of two PRs to remove the domain model before we launch

As discussed on slack, this PR strips out client side cmds for all
CRUD type operations in medic around orgs, roles, users , projects.

I also removed the old key generation code which we are not using
at present and have no immediate need to use.

Following this will be a PR to remove / disable the equivalent APIs

Signed-off-by: Luke Hinds <luke@stacklok.com>