engine: Pass params to the remediate interface, too
In order to support the branch protection remediations, we need to have
access to the parameters in the remediator. Those include typically the
branch name in the case of branch protection rules.
For the branch protection remediations, we need to pass in the params as
well in order to expand the branch name.