Add PRs to the DB in a migration
Initially we stopped half-way when implementing PRs in the sense that
the engine handled them, but we didn't store the PRs in the database.
This worked fine until we added alert processing which needs the
previous state of the entity being evaluated.

Instead of adding even more special casing for PRs, it's cleaner to just
store PRs in the database.