feat: Implement ingester cache
This implements a cache which stores the values of ingester calls. This is useful
so we may re-use the values of previous calls without having to hammer providers
as we do today.

The cache keys are computed with a sha-512 checksum which uses the following
inputs:

* ingester type
* ingester configuration (marshaled, if available)
* entity (marshaled)
* rule parameters (marshaled, if available)

We store the `RuleResult` pointer and re-use it on subsequent calls.

The lifetime of the cache is merely an executor entity evaluation. This means,
These calls will be re-used on one or multiple profiles for one round of
execution. this ensures that the next event we get for an entity is not cached, and
we get fresh values for it.