Skip non-relevant webhook events
This patch makes mediator more strict when it comes to mapping webhook
events, as not all events should trigger entity evaluation. Previously,
we have been using repo evaluation as sort of a fallback, but in busy
repositories with many PRs and issues, this could cause mediator to just
spin all the time.

Fixes: #1267