feat: Make watermill settings configurable
This adds configuration options for our watermill implementation. For now we only
have a `go-channel` driver, which is what we've been using until now. This is
also the default.

The first configuration settings this adds for the aforementioned driver as the following:

* `router_close_timeout`: per the watermill docs this determines how long router should work for handlers when closing.
* `buffer_size`: determines the buffer size for the golang channel the driver is using.

By allowing for setting the buffer size, we allow for paralellization of
request handling of events. We were not setting this before and thus
our event handling was serial.

We should also configure the `router_close_timeout` to be slightly lower than
the pod's termination grace period in our deployment. A separate PR will make
that setting configurable.