feat: Add possibility for entitlements
Entitlements are a set of features that are enabled in projects.

For this, we have a `features` table which represents the features available
in the system. These features have each a some settings that control
how they behave.

Features are then linked to projects via the `entitlements` table.

There is a helper function called `GetFeatureInProject` that allows us
to verify if a feature is enabled for a certain project.

Note that there is no way of defining a feature or setting an entitlement
via code. This has to be an administrator action via a direct database connection.