Add rule type validation on updates
This gives us safer rule type updates by adding a validation step
before doing the actual update.

If there are profiles using the rule, it'll run the validation, else it'll
just safely update the rule.

The validation checks that we don't introduce changes that would break the
schemas: either the rule schema itself or the parameters.