Run minder in a read-only filesystem (#1422)
This changes the helm chart to configure the minder `Deployment` to use a read-only
root filesystem. This helps us harden the deployment and makes it harder for an
attacker to keep a persistent rootkit in the pod.

Note that the docker-compose set up has already been using this setting to
ensure we don't break minder.