Fix flaky database tests (#1588)
We used to silently delete the repository if it happened that we chose
the same "repo id" when creating a random repository. This is not correct
and lead to hard-to-find and flaky tests.

Now, we instead check if the repo id exists and if it does, we generate a new ID
before creating the repo.