Enable skipping teardown when using the `run-docker` Makefile target
Now you can use an environment variable to tell our `run-docker` target
to skip the teardown phase. It may be done as follows:

```bash
export RUN_DOCKER_NO_TEARDOWN=true
```

This will result in subsequent calls to `make teardown` to skip the
`docker-compose down` call at the beginning of the run.

this is handy for cases where you're doing rapid development and want
to keep the database changes.