Add switch to migrate up/down to do one step migration, add migrate version (#1614)
When testing more complex database migrations, it's useful to do
migration one step up or down. This patch adds a new parameter to
migrate up and down that allows to specify the number of steps,
defaulting to migrating the whole way up or down.

Also adds a new migrate subcommand version that prints the current
database version.