Handle PR statuses when setting branch protection rules
The GH API returns both `Protection.RequiredStatusChecks.Contexts` which
is deprecated and `Protection.RequiredStatusChecks.Checks` which is
supposed to be used now on GET, but the PUT update needs to only set one
or the other.

Let's filter out the deprecated field so that setting branch protection
rules when commit statuses are in place works fine.

Fixes: #1596